Small fix of dtype propagation in Nose-Hoover thermostat.

+++ jax_md/simulate.py
@@ -257,7 +257,7 @@ def nvt_nose_hoover(
 
     DOF, = static_cast(R.shape[0] * R.shape[1])
 
-    Q = T * tau ** f32(2) * np.ones(chain_length)
+    Q = T * tau ** f32(2) * np.ones(chain_length, dtype=R.dtype)
     Q = ops.index_update(Q, 0, Q[0] * DOF)
 
     KE, V, xi, v_xi = step_chain(KE, V, xi, v_xi, Q, DOF, T)
